Task/security

+++ src/main/java/com/vention/stockmarket/service/impl/SecurityServiceImpl.java
@@ -2,9 +2,12 @@ package com.vention.stockmarket.service.impl;
 
 import com.vention.stockmarket.domain.SecurityCredentials;
 import com.vention.stockmarket.dto.request.AuthRequestDTO;
+import com.vention.stockmarket.dto.request.RolesUpdateDTO;
 import com.vention.stockmarket.dto.response.ResponseDTO;
 import com.vention.stockmarket.exceptions.CustomResourceNotFoundException;
+import com.vention.stockmarket.exceptions.CustomUnauthorizedException;
 import com.vention.stockmarket.repository.SecurityRepository;
+import com.vention.stockmarket.service.SecurityHelperService;
 import com.vention.stockmarket.service.SecurityService;
 import com.vention.stockmarket.utils.JwtUtils;
 import lombok.RequiredArgsConstructor;
@@ -20,6 +23,7 @@ public class SecurityServiceImpl implements SecurityService {
     private final SecurityRepository repository;
     private final PasswordEncoder passwordEncoder;
     private final JwtUtils jwtUtils;
+    private final SecurityHelperService securityHelperService;
 
     @Override
     public ResponseDTO<Long> create(SecurityCredentials securityCredentials) {
@@ -33,22 +37,38 @@ public class SecurityServiceImpl implements SecurityService {
         if (securityModel.isEmpty()) {
             throw new CustomResourceNotFoundException("Security model not found with id : " + id);
         }
-
+        if (!SecurityHelperService.hasUserPermissions(securityModel.get().getEmail())) {
+            throw new CustomUnauthorizedException();
+        }
         return new ResponseDTO<>(true, securityModel.get());
     }
 
     @Override
     public void update(SecurityCredentials securityCredentials) {
+        if (!SecurityHelperService.hasUserPermissions(securityCredentials.getEmail())) {
+            throw new CustomUnauthorizedException();
+        }
+        securityCredentials.setPassword(passwordEncoder.encode(securityCredentials.getPassword()));
         repository.update(securityCredentials);
     }
 
     @Override
     public void delete(Long id) {
+        var securityModel = repository.getById(id);
+        if (securityModel.isEmpty()) {
+            throw new CustomResourceNotFoundException("Security model not found with id : " + id);
+        }
+        if (!SecurityHelperService.hasUserPermissions(securityModel.get().getEmail())) {
+            throw new CustomUnauthorizedException();
+        }
         repository.delete(id);
     }
 
     @Override
     public ResponseDTO<List<SecurityCredentials>> getAll() {
+        if (!SecurityHelperService.isAdmin()) {
+            throw new CustomUnauthorizedException();
+        }
         var all = repository.getAll();
         return new ResponseDTO<>(true, all);
     }
@@ -61,5 +81,26 @@ public class SecurityServiceImpl implements SecurityService {
             return jwtUtils.generateJwt(authRequest.getEmail());
         }
         return "";
+    }
+
+    @Override
+    public ResponseDTO<SecurityCredentials> getByEmail(String email) {
+        if (!SecurityHelperService.hasUserPermissions(email)) {
+            throw new CustomUnauthorizedException();
+        }
+        var securityCredentials = repository.getByEmail(email);
+        if (securityCredentials.isEmpty()) {
+            throw new CustomResourceNotFoundException("security credentials not found with this email " + email);
+        }
+        return new ResponseDTO<>(true, securityCredentials.get());
+    }
+
+    @Override
+    public ResponseDTO<?> editRoles(RolesUpdateDTO updateDTO) {
+        if (!SecurityHelperService.isAdmin()) {
+            throw new CustomUnauthorizedException();
+        }
+        repository.updateRoles(updateDTO);
+        return new ResponseDTO<>(true, 200);
     }
 }
